Exception handling - delete

+++ src/main/java/com/educandoweb/coursejava/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.coursejava.entities.User;
 import com.educandoweb.coursejava.repositories.UserRepository;
+import com.educandoweb.coursejava.services.exceptions.DataBaseException;
 import com.educandoweb.coursejava.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -19,20 +21,37 @@ public class UserService {
 	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
+
+	// modern implementation
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseException(e.getMessage());
+		}
 	}
-	
+
+//		try {
+//			repository.deleteById(id);
+//		} catch (EmptyResultDataAccessException e) {
+//			throw new ResourceNotFoundException(id);
+//		} catch (DataIntegrityViolationException e) {
+//			throw new DatabaseException(e.getMessage());
+//		}
+//	}
+
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, obj);
@@ -42,6 +61,6 @@ public class UserService {
 	private void updateData(User entity, User obj) {
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
-		entity.setPhone(obj.getPhone());		
+		entity.setPhone(obj.getPhone());
 	}
 }
